requirements.txt
# Core dependencies for Scientific RAG System
arxiv==2.1.0
unstructured[pdf]==0.18.13
chromadb==0.4.22
sentence-transformers==2.2.2
transformers==4.36.2
torch>=2.6.0
streamlit==1.29.0
accelerate==0.25.0
datasets==2.16.1
numpy>=1.24.0
pandas>=2.1.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
huggingface-hub>=0.19.0
pillow>=10.1.0
pypdf>=3.17.0
pdf2image>=1.16.0
